refactor(frontend): rename sources-list to evidence-list
Align the module path with the EvidenceList export and update PostHog event documentation to match.
